requirements.txt
Django>=5.2.11
Pillow>=10.0.0
python-dotenv>=1.0.0
django-payments>=2.0.0
stripe>=5.0.0
